# frontend/src/routes/_layout/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Full Stack FastAPI Project", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/products/": {
      "get": {
        "tags": ["products"],
        "summary": "Read Products",
        "description": "Retrieve products.",
        "operationId": "products-read_products",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProductsPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["products"],
        "summary": "Create Product",
        "description": "Create new product.",
        "operationId": "products-create_product",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProductCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProductPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "tags": ["products"],
        "summary": "Read Product",
        "description": "Get product by ID.",
        "operationId": "products-read_product",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProductPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["products"],
        "summary": "Update Product",
        "description": "Update a product.",
        "operationId": "products-update_product",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProductUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProductPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["products"],
        "summary": "Delete Product",
        "description": "Delete a product.",
        "operationId": "products-delete_product",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { "type": "string" },
                  "title": "Response Products-Delete Product"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MediaPublic": {
        "properties": {
          "uuid": { "type": "string", "maxLength": 36, "title": "Uuid" },
          "file_name": {
            "type": "string",
            "maxLength": 255,
            "title": "File Name"
          },
          "mime_type": {
            "anyOf": [
              { "type": "string", "maxLength": 100 },
              { "type": "null" }
            ],
            "title": "Mime Type"
          },
          "size": { "type": "integer", "title": "Size" },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": ["uuid", "file_name", "size", "id", "created_at"],
        "title": "MediaPublic"
      },
...
      "ProductCategoryPublic": {
        "properties": {
          "name": { "type": "string", "maxLength": 255, "title": "Name" },
          "description": {
            "anyOf": [
              { "type": "string", "maxLength": 500 },
              { "type": "null" }
            ],
            "title": "Description"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": ["name", "id", "created_at"],
        "title": "ProductCategoryPublic"
      },
      "ProductCreate": {
        "properties": {
          "name": { "type": "string", "maxLength": 255, "title": "Name" },
          "description": {
            "anyOf": [
              { "type": "string", "maxLength": 1000 },
              { "type": "null" }
            ],
            "title": "Description"
          },
          "buying_price": {
            "anyOf": [
              { "type": "number" },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Buying Price"
          },
          "selling_price": {
            "anyOf": [
              { "type": "number", "exclusiveMinimum": 0.0 },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              }
            ],
            "title": "Selling Price"
          },
          "current_stock": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Current Stock",
            "default": 0
          },
          "reorder_level": {
            "anyOf": [
              { "type": "integer", "minimum": 0.0 },
              { "type": "null" }
            ],
            "title": "Reorder Level"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "status_id": {
            "type": "string",
            "format": "uuid",
            "title": "Status Id"
          },
          "image_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Image Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "buying_price",
          "selling_price",
          "category_id",
          "status_id"
        ],
        "title": "ProductCreate"
      },
      "ProductPublic": {
        "properties": {
          "name": { "type": "string", "maxLength": 255, "title": "Name" },
          "description": {
            "anyOf": [
              { "type": "string", "maxLength": 1000 },
              { "type": "null" }
            ],
            "title": "Description"
          },
          "buying_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$",
            "title": "Buying Price",
            "default": "0"
          },
          "selling_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$",
            "title": "Selling Price"
          },
          "current_stock": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Current Stock",
            "default": 0
          },
          "reorder_level": {
            "anyOf": [
              { "type": "integer", "minimum": 0.0 },
              { "type": "null" }
            ],
            "title": "Reorder Level"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "status_id": {
            "type": "string",
            "format": "uuid",
            "title": "Status Id"
          },
          "image_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Image Id"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "category": { "$ref": "#/components/schemas/ProductCategoryPublic" },
          "status": { "$ref": "#/components/schemas/ProductStatusPublic" },
          "image": {
            "anyOf": [
              { "$ref": "#/components/schemas/MediaPublic" },
              { "type": "null" }
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "selling_price",
          "category_id",
          "status_id",
          "id",
          "created_at",
          "category",
          "status",
          "image"
        ],
        "title": "ProductPublic"
      },
      "ProductStatusPublic": {
        "properties": {
          "name": { "type": "string", "maxLength": 100, "title": "Name" },
          "description": {
            "anyOf": [
              { "type": "string", "maxLength": 500 },
              { "type": "null" }
            ],
            "title": "Description"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" }
        },
        "type": "object",
        "required": ["name", "id"],
        "title": "ProductStatusPublic"
      },
      "ProductUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              { "type": "string", "maxLength": 255 },
              { "type": "null" }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              { "type": "string", "maxLength": 1000 },
              { "type": "null" }
            ],
            "title": "Description"
          },
          "buying_price": {
            "anyOf": [
              { "type": "number", "minimum": 0.0 },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              },
              { "type": "null" }
            ],
            "title": "Buying Price"
          },
          "selling_price": {
            "anyOf": [
              { "type": "number", "exclusiveMinimum": 0.0 },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              },
              { "type": "null" }
            ],
            "title": "Selling Price"
          },
          "current_stock": {
            "anyOf": [
              { "type": "integer", "minimum": 0.0 },
              { "type": "null" }
            ],
            "title": "Current Stock"
          },
          "reorder_level": {
            "anyOf": [
              { "type": "integer", "minimum": 0.0 },
              { "type": "null" }
            ],
            "title": "Reorder Level"
          },
          "category_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Category Id"
          },
          "status_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Status Id"
          },
          "image_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Image Id"
          }
        },
        "type": "object",
        "title": "ProductUpdate"
      },
      "ProductsPublic": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/ProductPublic" },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "ProductsPublic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" }
        }
      }
    }
  }
}